requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.11.0
gunicorn==21.2.0
drf-spectacular==0.27.0
django-cors-headers==4.3.1
